moved sort functions out and removed density sort
